requirements.txt
langchain
pandas
google-generativeai
python-dotenv
pandas
scikit-learn
streamlit
langchain_community
